python/pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "rmcs_api_client"
version = "0.3.1"
authors = [
  { name="Chandra Wijaya Sentosa", email="chandra.w.sentosa@gmail.com" },
]
description = "rmcs API client library"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "grpcio",
    "grpcio-tools"
]
